+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.4.0 (No ETA)
+* Add ranks
+* Add chat police rank
+* Add parameter in menu system
+* Add bindings in config file
+* Fix : Disconnection won't change the job id
+* Fix : Heading for police vehicles
+
 ## 1.3.0 (07/07/2017)
 * FXServer compatibility (drop mysql support on FXServer version)
 * New en translation (thanks BCTB and BritishBrotherhood) and de translation (thanks SyfuxX)